회원 가입 쿼리 작성
Contents
- 회원가입 쿼리 작성
- 회원 빌더 createAt추가

Todo
- 단일 회원 조회 쿼리 작성
- 전체 회원 조회(임시) 작성

+++ src/main/java/com/kakao/cafe/core/repository/member/MemberRepository.java
@@ -4,6 +4,6 @@ import com.kakao.cafe.core.domain.member.Member;
 
 import java.util.List;
 
-public interface MemberRepository extends CrudRepository<Member, Long> {
+public interface MemberRepository extends CrudRepository<Member, Integer> {
     List<Member> findAll();
 }
